speed up points-to calculation
in case we calculate SparseIntSet.containsAny for a MutableSharedBitVectorIntSet
that consists of two SparseIntSet, we currently calculate MutableSharedBitVectorIntSet.contains
for every element of the SparseIntSet. We now delegate the check to
MutableSharedBitVectorIntSet.containsAny(SparseIntSet) instead, which in turn calls containsAny on
its two parts. Then we get the much more efficient (linear) algorithm of
SparseIntSet.containsAny(SparseIntSet), which functions analogously to merging.