have to change the way an AndroidEntryPointManager is passed around in the manifest reader
The change is necessary since the *Item objects are created in a
context where the manager is not available. Hence, instead of passing
the manager to the constructor, it is passed as parameter in the
enter/leave methods now.

Furthermore, The *Item classes need to have a public constructor, since
they are instantiated via reflection.